updates based on pr feedback - remove unused id fields - renamed page and routes from individual events to my events to better match what is displayed to users - added margins to bottom of table for white space
