+++ app/javascript/pages/MyEvents/index.js
