+++ app/javascript/pages/MyEvents/main.css
@@ -156,3 +156,7 @@ textarea {
 .eventsTable > div > div:nth-child(1) > div:nth-child(1) {
   box-shadow: inset 0 -2px #eee;
 }
+
+.eventsTable {
+  margin-bottom: 20px;
+}
